Renamed CGAL_OGL_VBO_helper to VBOPolyhedron
